+++ src/debug-core/target.h
@@ -110,7 +110,7 @@ public:
 	virtual void read_sfr( uint8_t addr, uint8_t len, unsigned char *buf )=0;
 	virtual void read_sfr( uint8_t addr, uint8_t page, uint8_t len, unsigned char *buf )=0;
 	virtual void read_xdata( uint16_t addr, uint16_t len, unsigned char *buf )=0;
-	virtual void read_code( uint16_t addr, uint16_t len, unsigned char *buf )=0;
+	virtual void read_code( uint16_t addr, int len, unsigned char *buf )=0;
 	virtual uint16_t read_PC()=0;
 	
 	// memory writes
@@ -118,7 +118,7 @@ public:
 	virtual void write_sfr( uint8_t addr, uint8_t len, unsigned char *buf )=0;
 	virtual void write_sfr( uint8_t addr, uint8_t page,uint8_t len, unsigned char *buf );
 	virtual void write_xdata( uint16_t addr, uint16_t len, unsigned char *buf )=0;
-	virtual void write_code( uint16_t addr, uint16_t len, unsigned char *buf )=0;
+	virtual void write_code( uint16_t addr, int len, unsigned char *buf )=0;
 	virtual void write_PC( uint16_t addr )=0;
 
 	/** load an intel hex file into the target
